+++ Private/New-DelegatedAccessToken.ps1
@@ -42,9 +42,9 @@ function New-DelegatedAccessToken {
         [ValidateNotNullOrEmpty()]
         [String]$RedirectUri = "http://localhost",
 
-        [parameter(Mandatory = $false, HelpMessage = "Array of permission scopes to request.")]
+        [parameter(Mandatory = $true, HelpMessage = "Array of permission scopes to request.")]
         [ValidateNotNullOrEmpty()]
-        [String[]]$Scopes = @("DeviceManagementApps.ReadWrite.All", "DeviceManagementRBAC.Read.All", "offline_access")
+        [String[]]$Scopes
     )
     Process {
         try {